requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.8.0
scikit-image>=0.19.0
# 可选：仅当使用深度学习风格迁移时需要
# torch>=2.0.0
# torchvision>=0.15.0

